+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 'use strict'
 import { articlesByUuids, featureArticles, relatedArticles } from './articles'
-import { authorsList } from './authors'
+import { searchedAuthorsList, authorsList } from './authors'
 import { articlesByAuthor } from './author-articles'
 import { categories, tags } from './groups'
 import { combineReducers } from 'redux'
@@ -69,6 +69,7 @@ const rootReducer = combineReducers({
   header,
   articleSlugToId,
   topicSlugToId,
+  searchedAuthorsList,
   authorsList,
   articlesByAuthor,
   entities