+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ### Unreleased
 
 - [Update] Move author state formatting from action to reducer
+- [Update] Seperate authors state into searched authors and all authors
 - [Testing] API testing fully completed 
+
 
 ### 1.1.5
 - [Update] content of sub-nav-path